First public release
Have the basic grid down and a way to track the state of the grid, i
think. Right now it just puts random colors into the grid, and there is
some simple event handling for arrow keys. Soon life will live. Almost
everything is in place. Just need to figure out mouse handling, and and
to implement a rules algorithm, and an algorithm to take the state of
the game and draw it to the screen.